Remove keyword filter, use same component for sources
