+++ app/components/Sources/styles.css
@@ -1,17 +1,7 @@
-.sources-wrapper {
+.sources {
     display: flex;
-    align-items: center;
-    padding: var(--dui-spacing-small);
-    gap: var(--dui-spacing-small);
-
-    .label {
-        flex-grow: 1;
-        padding-left: var(--dui-spacing-small);
-    }
-
-    .info-icon {
-        flex-shrink: 0;
-        color: var(--color-text-title);
-        font-size: var(--dui-font-size-extra-large);
-    }
+    flex-direction: column;
+    background-color: transparent;
+    width: 40vw;
+    gap: var(--dui-spacing-extra-small);
 }
