+++ app/views/Dashboard/CombinedIndicators/styles.css
@@ -4,6 +4,16 @@
     justify-content: center;
     gap: var(--dui-spacing-extra-large);
 
+    .themes {
+        &.empty {
+            display: flex;
+            align-items: center;
+            flex-direction: column;
+            justify-content: center;
+            min-height: 40vh;
+        }
+    }
+
     .thematic-container {
         display: flex;
         flex-direction: column;
@@ -13,24 +23,5 @@
             background-color: transparent;
             gap: 0;
         }
-
-    }
-
-    .source-heading {
-        display: flex;
-        justify-content: space-between;
-        margin-top: var(--dui-spacing-medium);
-        padding: var(--dui-spacing-medium) 0px;
-        width: 40vw;
-        font-size: var(--dui-font-size-heading-small);
-        font-weight: var(--dui-font-weight-bold);
-        gap: var(--dui-spacing-small);
-    }
-
-    .sources {
-        display: flex;
-        flex-direction: column;
-        width: 40vw;
-        gap: var(--dui-spacing-small);
     }
 }
