+++ app/views/Dashboard/Country/styles.css
@@ -3,6 +3,10 @@
     flex-direction: column;
     gap: var(--dui-spacing-large);
 
+    .heading {
+        font-weight: var(--dui-font-weight-normal);
+    }
+
     .country-main {
         display: flex;
         gap: var(--dui-spacing-large);
@@ -79,10 +83,6 @@
                     border-radius: 0;
                     /* By default, ContainerCard has border-radius */
                     box-shadow: 0px 0px 10px var(--dui-card-box-shadow);
-
-                    .heading {
-                        font-weight: var(--dui-font-weight-normal);
-                    }
 
                     .gender-disaggregation-content {
                         display: flex;
@@ -182,12 +182,5 @@
                 }
             }
         }
-    }
-
-    .sources {
-        display: flex;
-        flex-direction: column;
-        gap: var(--dui-spacing-small);
-        max-width: 50rem;
     }
 }
