+++ package.json
@@ -51,7 +51,6 @@
         "react-dom": "^17.0.2",
         "react-ga": "^3.3.0",
         "react-icons": "^4.3.1",
-        "react-select": "^5.4.0",
         "react-slider": "^2.0.0",
         "react-transition-group": "^4.4.5",
         "recharts": "^2.1.9",
